c90 adjustment

+++ more_malloc_free/101-mul.c
@@ -9,12 +9,12 @@
 
 int main(int num1, char **num2)
 {
-	if (num1 != 2 || atoi(num2[0]) == 0 || atoi(num2[1]) == 0)
+	if (num1 != 2 || atol(num2[0]) == 0 || atol(num2[1]) == 0)
 	{
 		printf("Error\n");
 		return (98);
 	}
 
-	printf("%lld\n", (long long)(atoi(num2[0])) * (long long)(atoi(num2[1])));
+	printf("%ld\n", atol(num2[0]) * atol(num2[1]));
 	return (0);
 }
